pyproject.toml
[tool.poetry]
name = "fastapi-starter"
version = "0.1.0"
description = "FastAPI starter project"
authors = ["Silas_229 <contact@silas229.de>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
sqlmodel = "^0.0.22"
uvicorn = "^0.32.0"
gunicorn = "^23.0.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = { extras = ["email"], version = "^2.9.2" }
pydantic-settings = "^2.6.0"
python-multipart = "^0.0.12"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
